requirements.txt
aiogram
openpyxl
google-genai
flask
psycopg2-binary
openpyxl